[3.4.1] - Big Organisation update
- all UI functions move to UI_functions.py
- Save function move to SavingFunctions.py
- Save battle muted
