+++ BattleWindow.py
@@ -1,15 +1,13 @@
 import tkinter as tk
 from tkinter import ttk
-import json
-import os
-from Classes import Division,Camp
-from Functions.TacticsFunctions import choose_tactic, change_weight
+from Classes.Camp import Camp
+from Functions.TacticsFunctions import choose_tactic
+from Functions.UI_functions import *
+from Functions.SavingFunctions import *
 from Library.TerrainList import terrain_list
 from Library import LeaderList
 from Library.LeaderList import *
 
-Division = Division.Division # shortcut
-Camp = Camp.Camp # shortcut
 
 class BattleWindow(tk.Tk):
     def __init__(self):
@@ -34,7 +32,7 @@ class BattleWindow(tk.Tk):
         self.camp_defender.is_attacking = False
 
         # Charger les divisions sauvegardées
-        self.divisions = self.load_divisions()
+        self.divisions = load_divisions(self)
 
         # Cadre pour les paramètres de la bataille
         frame_params = tk.Frame(self)
@@ -48,13 +46,13 @@ class BattleWindow(tk.Tk):
         self.attacker_leader_label = tk.Label(self.frame_leaders, text="Leader Attaquant : Aucun",
                                               font=("Arial", 12, "bold"), bd=2, relief=tk.RIDGE)
         self.attacker_leader_label.pack(side=tk.LEFT, padx=20)
-        self.attacker_leader_label.bind("<Button-1>", lambda event: self.select_leader("attacker"))
+        self.attacker_leader_label.bind("<Button-1>", lambda event: select_leader(self,"attacker"))
 
         # Label pour le leader du camp défenseur
         self.defender_leader_label = tk.Label(self.frame_leaders, text="Leader Défenseur : Aucun",
                                               font=("Arial", 12, "bold"), bd=2, relief=tk.RIDGE)
         self.defender_leader_label.pack(side=tk.RIGHT, padx=20)
-        self.defender_leader_label.bind("<Button-1>", lambda event: self.select_leader("defender"))
+        self.defender_leader_label.bind("<Button-1>", lambda event: select_leader(self,"defender"))
 
         # Ajout des labels pour les tactiques et la phase de bataille
         self.battle_phase_label = tk.Label(self, text="Phase de Bataille : Aucune", font=("Arial", 12, "bold"))
@@ -85,19 +83,19 @@ class BattleWindow(tk.Tk):
         terrain_names = [terrain.name for terrain in terrain_list]
         self.terrain_dropdown = ttk.Combobox(frame_params, textvariable=self.selected_terrain, values=terrain_names)
         self.terrain_dropdown.grid(row=0, column=3, padx=5)
-        self.terrain_dropdown.bind("<<ComboboxSelected>>", self.update_terrain)
+        self.combat_width_display = tk.StringVar()
+        self.combat_width = terrain_list[0].width
+        self.terrain_dropdown.bind("<<ComboboxSelected>>", update_terrain(self))
 
         # Ajouter les cases à cocher pour les rivières
         tk.Checkbutton(frame_params, text="Petite Rivière", variable=self.small_river_box,
-                       command=self.on_river_change).grid(row=0, column=4, padx=5)
+                       command=on_river_change(self)).grid(row=0, column=4, padx=5)
         tk.Checkbutton(frame_params, text="Grande Rivière", variable=self.large_river_box,
-                       command=self.on_river_change).grid(row=1, column=4, padx=5)
+                       command=on_river_change(self)).grid(row=1, column=4, padx=5)
         tk.Checkbutton(frame_params, text="Encerclement", variable=self.encirclement_box,
-                       command=self.on_encirclement_change).grid(row=0, column=5, padx=5)
+                       command=on_encirclement_change(self)).grid(row=0, column=5, padx=5)
 
-        self.combat_width = terrain_list[0].width
         tk.Label(frame_params, text="Aire de combat").grid(row=2, column=0, padx=5)
-        self.combat_width_display = tk.StringVar()
         self.combat_width_display.set(str(self.combat_width))  # Set default width
         tk.Label(frame_params, textvariable=self.combat_width_display).grid(row=2, column=1, padx=5)
 
@@ -123,7 +121,7 @@ class BattleWindow(tk.Tk):
         self.camp_attacker_divisions_frame = tk.Frame(frame_camp_attacker)
         self.camp_attacker_divisions_frame.pack(fill=tk.BOTH, expand=True, pady=5)
 
-        tk.Button(frame_camp_attacker, text="Ajouter Division", command=lambda: self.open_division_selection(self.camp_attacker, self.camp_attacker_divisions_frame)).pack()
+        tk.Button(frame_camp_attacker, text="Ajouter Division", command=lambda: open_division_selection(self,self.camp_attacker, self.camp_attacker_divisions_frame)).pack()
 
         # Camp B
         frame_camp_defender = tk.Frame(frame_battle)
@@ -133,11 +131,11 @@ class BattleWindow(tk.Tk):
         self.camp_defender_divisions_frame = tk.Frame(frame_camp_defender)
         self.camp_defender_divisions_frame.pack(fill=tk.BOTH, expand=True, pady=5)
 
-        tk.Button(frame_camp_defender, text="Ajouter Division", command=lambda: self.open_division_selection(self.camp_defender, self.camp_defender_divisions_frame)).pack()
+        tk.Button(frame_camp_defender, text="Ajouter Division", command=lambda: open_division_selection(self,self.camp_defender, self.camp_defender_divisions_frame)).pack()
 
         # Bouton pour lancer un round de la bataille
         tk.Button(self, text="Lancer un Round", command=self.run_battle_round).pack(pady=10)
-        tk.Button(self, text="Sauvegarder Bataille", command=self.save_battle_data).pack(pady=10)
+        #tk.Button(self, text="Sauvegarder Bataille", command=save_battle_data(self)).pack(pady=10)
 
         # Zone de logs pour les résultats
         self.log_text = tk.Text(self, height=10)
@@ -181,9 +179,9 @@ class BattleWindow(tk.Tk):
         self.round_counter += 1
 
         # Mets a jour affichage
-        self.refresh_display()
-        self.update_battle_info_display()
-        self.update_combat_width()
+        refresh_display(self)
+        update_battle_info_display(self)
+        update_combat_width(self)
 
 
     def _round(self,camp_attacker,camp_defender):
@@ -220,7 +218,6 @@ class BattleWindow(tk.Tk):
                 if division.pv <= 0 or division.organisation <= 0:
                     self.retreat_division(division)
 
-    ###################################################
     def retreat_division(self,division_to_retreat):
         """
         Retire la division de toute les listes de la bataille. Pour représenté que la division s'est replié du champ de
@@ -246,232 +243,12 @@ class BattleWindow(tk.Tk):
                         widget.destroy()  # Détruire le cadre correspondant à la division retirée
                         break
 
-    def update_terrain(self,event=None):
-        """
-        Recupère l'instance :terrain: a partir du choix fait dans le menu déroulant.
-        :return:
-        """
-        selected_terrain_name = self.selected_terrain.get()
-        for terrain in terrain_list:
-            if terrain.name == selected_terrain_name:
-                self.terrain = terrain
-                break
-        self.update_combat_width()
 
-    def update_combat_width(self):
-        """
-        Mets à jour l'affichage du combat_width
-        """
-        self.combat_width = self.terrain.width
-        try:
-            self.combat_width *= self.camp_attacker.tactic.width_bonus * self.camp_defender.tactic.width_bonus
-        except Exception:
-            pass
-        self.combat_width_display.set(str(self.combat_width))
-
-    def update_battle_info_display(self):
-        """
-        Met à jour l'affichage des tactiques pour chaque camp et la phase de bataille en cours.
-        """
-        self.attacker_tactic_label.config(text=f"Tactique Attaquant : {self.camp_attacker.tactic.name}")
-        self.defender_tactic_label.config(text=f"Tactique Défenseur : {self.camp_defender.tactic.name}")
-        self.battle_phase_label.config(text=f"Phase de Bataille : {self.battle_phase}")
 
 
     ############# BOUTONS ####################
 
-    def load_divisions(self):
-        """
-           Charge les divisions sauvegardées à partir d'un fichier JSON.
-           Cette méthode vérifie l'existence du fichier "divisions.json" et tente de charger les données JSON à partir
-           de ce fichier.
-           Les données sont ensuite converties en instances de la classe Division.
-           Retour:
-               List[Division]: Une liste d'instances de la classe Division représentant les divisions sauvegardées.
-           """
-        if os.path.exists("Saves/divisions.json"):
-            with open("Saves/divisions.json", "r") as file:
-                try:
-                    data = json.load(file)
-                    return [Division.load(division) for division in data]
-                except json.JSONDecodeError:
-                    return []
-        return []
 
-    def get_division_names(self):
-        """
-           Récupère les noms de toutes les divisions sauvegardées.
-           Cette méthode parcourt la liste des divisions chargées et extrait le nom de chaque division.
-           Retour:
-               List[str]: Une liste contenant les noms de toutes les divisions sauvegardées.
-           """
-        return [division.template for division in self.divisions]
-
-    def add_division(self, frame, selected_division_var):
-        """
-           Ajoute une division au camp spécifié et affiche ses statistiques de manière compacte.
-
-           Cette méthode recherche la division correspondant au nom sélectionné, crée un cadre pour la division,
-           et affiche ses statistiques sous forme abrégée en colonnes, avec des barres de progression pour les PV et l'organisation.
-
-           Args:
-               frame (tk.Frame): Le cadre dans lequel ajouter la division (camp A ou camp B).
-               selected_name (str): Le Nom de Template sélectionnée à ajouter.
-           """
-        #todo Ajouter les division directement dans la réserves via les bouttons
-        if not (selected_name := selected_division_var.get()):
-            return
-        for division in self.divisions:
-            if division.template == selected_name:
-                camp = None
-                if frame == self.camp_attacker_divisions_frame:
-                    self.camp_attacker.add_division(division)
-                    camp = self.camp_attacker
-                if frame == self.camp_defender_divisions_frame:
-                    self.camp_defender.add_division(division)
-                    camp = self.camp_defender
-                assert camp is not None , "camp is not assigned"
-                frame_division = tk.Frame(frame, bd=1, relief=tk.SOLID, padx=5, pady=5)
-                frame_division.pack(fill=tk.X, pady=2)
-
-                stats_frame = tk.Frame(frame_division)
-                stats_frame.pack(fill=tk.X)
-
-                self.display_division_stats(stats_frame,division,camp)
-
-    def get_divisions_from_frame(self, frame):
-        """
-           Récupère les divisions à partir d'un cadre spécifié.
-           Cette méthode extrait les noms des divisions à partir des enfants du cadre donné,
-           puis recherche et retourne les instances de Division correspondantes.
-           Args:
-               frame (tk.Frame): Le cadre contenant les divisions.
-           Retour:
-               List[Division]: Une liste d'instances de la classe Division correspondant aux noms dans le cadre.
-           """
-        division_names = [child.winfo_children()[0].cget("text").split(":")[1].strip() for child in
-                          frame.winfo_children()]
-        return [
-            next(
-                (division for division in self.divisions if division.nom == name),
-                None,
-            )
-            for name in division_names
-        ]
-
-    def open_division_selection(self, camp, frame):
-        selection_window = tk.Toplevel(self)
-        selection_window.title("Sélectionner une Division")
-        selection_window.geometry("400x300")
-
-        listbox = tk.Listbox(selection_window)
-        listbox.pack(fill=tk.BOTH, expand=True)
-
-        for division in self.divisions:
-            listbox.insert(tk.END, division.template)
-
-        def on_select():
-            selected_name = tk.StringVar()
-            selected_name.set(listbox.get(listbox.curselection()))
-            self.add_division( frame, selected_name)
-            #selection_window.destroy()
-
-        tk.Button(selection_window, text="Ajouter", command=on_select).pack(pady=10)
-
-    def save_battle_data(self):
-        battle_data = {
-            "weather": self.weather.get(),
-            "terrain": self.terrain.get(),
-            "leader_a": self.leader_attacker.get(),
-            "leader_b": self.leader_defender.get(),
-            "camp_a": self.camp_attacker.get_data(),
-            "camp_b": self.camp_defender.get_data(),
-            "log_text": self.log_text.get("1.0", tk.END).strip()
-        }
-
-        with open("Saves/battle_data.json", "w") as file:
-            json.dump(battle_data, file, indent=4)
-
-    def display_division_stats(self,stats_frame,division,camp):
-        # Déterminer la couleur en fonction de la position de la division
-        if division in camp.frontline:
-            bg_color = "lightblue"  # Couleur pour les divisions en frontline
-        elif division in camp.reserves:
-            bg_color = "lightgray"  # Couleur pour les divisions en réserves
-        else:
-            bg_color = "white"  # Couleur par défaut si la division n'est ni en frontline ni en réserves
-
-        # Appliquer la couleur de fond
-        stats_frame.configure(bg=bg_color)
-
-        stats = [division.pv, division.organisation, division.soft_attack, division.hard_attack, division.defense,
-                 division.attaque, division.piercing, division.armor, division.hardness, division.width]
-        abbr_stats = ["PV", "Org", "SA", "HA", "Def", "Atk", "Prc", "Arm", "Hard", "Wdth"]
-
-        for i, stat in enumerate(stats):
-            row = i // 6
-            col = i % 6
-            if stat in [division.pv, division.organisation]:
-                #FIXME -  TRUE lorsque :stat: à la meme valeur que :pv: ou :org:
-                tk.Label(stats_frame, text=f"{abbr_stats[i]}: {stat}").grid(row=row * 2, column=col)
-                value = stat
-                if stat == division.pv: max_value = division._PV
-                if stat == division.organisation: max_value = division._ORGANISATION
-                progress = ttk.Progressbar(stats_frame, maximum=max_value, value=value, length=80)
-                progress.grid(row=row * 2 + 1, column=col)
-            else:
-                tk.Label(stats_frame, text=f"{abbr_stats[i]}: {stat}").grid(row=row * 2, column=col)
-
-    def refresh_display(self):
-        for frame, current_camp in [(self.camp_attacker_divisions_frame, self.camp_attacker),
-                                    (self.camp_defender_divisions_frame, self.camp_defender)]:
-            for widget in frame.winfo_children():
-                widget.destroy()  # Effacer les anciennes stats
-
-            for division in current_camp.get_divisions():
-                division_frame = tk.Frame(frame, bd=1, relief=tk.SOLID, padx=5, pady=5)
-                division_frame.pack(fill=tk.X, pady=2)
-                stats_frame = tk.Frame(division_frame)
-                stats_frame.pack(fill=tk.X)
-                self.display_division_stats(stats_frame, division, current_camp)
-
-    def select_leader(self, camp_type):
-        leader_window = tk.Toplevel(self)
-        leader_window.title("Sélectionner un Leader")
-        leader_window.geometry("300x200")
-
-        listbox = tk.Listbox(leader_window)
-        listbox.pack(fill=tk.BOTH, expand=True)
-
-        # Remplir la liste avec les noms des leaders disponibles
-        leaders = self.get_leaders()
-        for leader in leaders:
-            listbox.insert(tk.END, leader.name)  # Assurez-vous que les instances Leader ont un attribut `name`
-
-        def on_select():
-            selected_name = listbox.get(listbox.curselection())
-            if camp_type == "attacker":
-                self.attacker_leader_label.config(text=f"Leader Attaquant : {selected_name}")
-                self.camp_attacker.add_leader(next((leader for leader in self.get_leaders() if leader.name == selected_name), None))
-            else:
-                self.defender_leader_label.config(text=f"Leader Défenseur : {selected_name}")
-                self.camp_defender.add_leader(next((leader for leader in self.get_leaders() if leader.name == selected_name), None))
-            leader_window.destroy()
-
-        tk.Button(leader_window, text="Sélectionner", command=on_select).pack(pady=10)
-
-    def get_leaders(self):
-        """
-        Retourne la liste des leaders disponibles pour le camp spécifié.
-        """
-        return [leader_A, leader_B, leader_C, leader_D, no_leader]
-
-    def on_river_change(self):
-        self.terrain.has_small_river = self.small_river_box.get()
-        self.terrain.has_large_river = self.large_river_box.get()
-
-    def on_encirclement_change(self):
-        self.encirclement = self.encirclement_box.get()
 
 app = BattleWindow()
 
@@ -481,17 +258,17 @@ if __name__ == "__main__" and test_case == "Case 1":
     app.attacker_leader_label.config(text=f"Leader Attaquant : {app.camp_attacker.leader.name}")
     app.camp_defender.add_leader(LeaderList.no_leader)
     app.defender_leader_label.config(text=f"Leader Defenseur : {app.camp_defender.leader.name}")
-    app.add_division(app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_attacker.get_divisions()[-1].nom = "Div. A1"
-    app.add_division(app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_attacker.get_divisions()[-1].nom = "Div. A2"
-    app.add_division(app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_attacker.get_divisions()[-1].nom = "Div. A3"
-    app.add_division(app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_defender.get_divisions()[-1].nom = "Div. B1"
-    app.add_division(app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_defender.get_divisions()[-1].nom = "Div. B2"
-    app.add_division(app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_defender.get_divisions()[-1].nom = "Div. B3"
 
 if __name__ == "__main__" and test_case == "Case 2":
@@ -499,8 +276,8 @@ if __name__ == "__main__" and test_case == "Case 2":
     app.attacker_leader_label.config(text=f"Leader Attaquant : {app.camp_attacker.leader.name}")
     app.camp_defender.add_leader(LeaderList.no_leader)
     app.defender_leader_label.config(text=f"Leader Defenseur : {app.camp_defender.leader.name}")
-    app.add_division(app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_attacker_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_attacker.get_divisions()[-1].nom = "Div. A"
-    app.add_division(app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
+    add_division(app,app.camp_defender_divisions_frame,tk.StringVar(value="Infanterie 36"))
     app.camp_attacker.get_divisions()[-1].nom = "Div. B"
 app.mainloop()
